+++ samples/unity-plugin/Assets/Bhaptics/SDK/AndroidWidget/Scripts/AndroidWidget_ScannedDeviceUI.cs
@@ -5,7 +5,6 @@ using UnityEngine.UI;
 
 namespace Bhaptics.Tact.Unity
 {
-
     public class AndroidWidget_ScannedDeviceUI : MonoBehaviour
     { 
         [Header("[UI]")]