+++ samples/unity-plugin/Assets/Bhaptics/SDK/Scripts/BhapticsUtils.cs
@@ -227,6 +227,58 @@ namespace Bhaptics.Tact.Unity
             return project;
         }
 
+        public static PositionType ToPositionType(HapticClipPositionType pos)
+        {
+            switch (pos)
+            {
+                case HapticClipPositionType.Head:
+                    return PositionType.Head;
+                case HapticClipPositionType.VestFront:
+                    return PositionType.VestFront;
+                case HapticClipPositionType.VestBack:
+                    return PositionType.VestBack;
+                case HapticClipPositionType.LeftHand:
+                    return PositionType.HandL;
+                case HapticClipPositionType.RightHand:
+                    return PositionType.HandR;
+                case HapticClipPositionType.LeftFoot:
+                    return PositionType.FootL;
+                case HapticClipPositionType.RightFoot:
+                    return PositionType.FootR;
+                case HapticClipPositionType.RightForearm:
+                    return PositionType.ForearmR;
+                case HapticClipPositionType.LeftForearm:
+                    return PositionType.ForearmL;
+            }
+
+            return PositionType.Head;
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="pos"></param>
+        /// <param name="isLeft">Value used for devices with left and right sides.(Default: Left)</param>
+        /// <returns></returns>
+        public static PositionType ToPositionType(HapticDeviceType pos, bool isLeft = true)
+        {
+            switch (pos)
+            {
+                case HapticDeviceType.Tactal:
+                    return PositionType.Head;
+                case HapticDeviceType.TactSuit:
+                    return PositionType.Vest;
+                case HapticDeviceType.Tactosy_arms:
+                    return isLeft ? PositionType.ForearmL : PositionType.ForearmR;
+                case HapticDeviceType.Tactosy_feet:
+                    return isLeft ? PositionType.FootL : PositionType.FootR;
+                case HapticDeviceType.Tactosy_hands:
+                    return isLeft ? PositionType.HandL : PositionType.HandR;
+            }
+
+            return PositionType.Head;
+        }
+
         public const string TypeHead = "Head";
         public const string TypeTactal = "Tactal";
         public const string TypeVest = "Vest";